add lib/processing/font.rb
